Filter background depth based on value and mask
Previously background depth was removed only based on segmask. However,
this left few pixels in the depth map that caused a large imbalance in
min-max range. Hence, now the background depth are also removed based on
a hardcoded value that corresponds to the scene settings when the data
was generated.